[fix/#156] : modal 위치 변경
